[adapters] Separate journal and checkpoint code in the controller.
These are related, but distinct, and it now seems worthwhile to separate
them.
