+++ crates/adapterlib/src/errors/journal.rs
