+++ crates/adapterlib/src/errors/mod.rs
@@ -1,2 +1,2 @@
 pub mod controller;
-pub mod metadata;
+pub mod journal;
